requirements.txt
python-telegram-bot==20.7
anthropic==0.21.3
openai==1.14.0
apscheduler==3.10.4
requests==2.31.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.5
